+++ .env.example
@@ -1,0 +1,15 @@
+# Django 설정
+DEBUG=True
+SECRET_KEY=django-insecure-your-secret-key-here-change-this-in-production
+
+# PostgreSQL 데이터베이스 설정
+# DB_USER는 본인의 macOS 사용자명을 입력 (터미널에서 'whoami' 명령어로 확인)
+DB_ENGINE=django.db.backends.postgresql
+DB_NAME=landing_project_db
+DB_USER=your-username-here
+DB_PASSWORD=
+DB_HOST=localhost
+DB_PORT=5432
+
+# 허용된 호스트
+ALLOWED_HOSTS=localhost,127.0.0.1,*
