+++ app/analysis/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class AnalysisConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'analysis'
+    verbose_name = '분석'
