+++ app/analysis/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import NotificationViewSet
+from .views import AnalysisViewSet
 
 router = DefaultRouter()
-router.register(r'notifications', NotificationViewSet, basename='notification')
+router.register(r'analysis', AnalysisViewSet, basename='analysis')
 
 urlpatterns = [
     path('', include(router.urls)),
