+++ app/notification/apps.py
@@ -1,9 +1,10 @@
 from django.apps import AppConfig
 
-class NotificationsConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "app.notifications"
+
+class NotificationConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'notification'
+    verbose_name = '알림'
 
     def ready(self):
-        # 시그널 등록
-        from . import signals  # noqa: F401
+        import notification.signals
