+++ app/notification/serializers.py
@@ -1,23 +1,12 @@
 from rest_framework import serializers
 from .models import Notification
 
+
 class NotificationSerializer(serializers.ModelSerializer):
     class Meta:
         model = Notification
         fields = [
-            "id",
-            "user",
-            "message",
-            "type",
-            "related_id",
-            "is_read",
-            "created_at",
-            "updated_at",
+            'id', 'message', 'notification_type', 'is_read',
+            'read_at', 'created_at'
         ]
-        read_only_fields = ["id", "user", "created_at", "updated_at"]
-
-class NotificationReadSerializer(serializers.Serializer):
-    ids = serializers.ListField(
-        child=serializers.IntegerField(min_value=1), required=False
-    )
-    all = serializers.BooleanField(required=False, default=False)
+        read_only_fields = ['id', 'read_at', 'created_at']
