+++ config/settings/base.py
@@ -1,9 +1,14 @@
 # 공통환경
 from pathlib import Path
+import os
+from dotenv import load_dotenv
+
+# .env 파일 로드
+load_dotenv()
 
 BASE_DIR = Path(__file__).resolve().parent.parent.parent
 
-SECRET_KEY = 'django-insecure-b3k(+600_pauf*u18ibuv!($*efb+wv+var$bk!6)cym%ve@4r'
+SECRET_KEY = os.getenv('SECRET_KEY', 'django-insecure-b3k(+600_pauf*u18ibuv!($*efb+wv+var$bk!6)cym%ve@4r')
 
 INSTALLED_APPS = [
     'django.contrib.admin',
@@ -12,12 +17,11 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'rest_framework',  # 추가
+    'rest_framework',
     'users',  # 사용자 앱
-    'app',
-    'app.accounts',
-    'app.accounts.constants',
-    'app.analysis',
+    'accounts',  # 계좌 관리 앱
+    'analysis',  # 분석 앱
+    'notification',  # 알림 앱
 ]
 
 MIDDLEWARE = [
@@ -51,8 +55,12 @@ WSGI_APPLICATION = 'config.wsgi.application'
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': BASE_DIR / 'db.sqlite3',
+        'ENGINE': os.getenv('DB_ENGINE', 'django.db.backends.postgresql'),
+        'NAME': os.getenv('DB_NAME', 'landing_project_db'),
+        'USER': os.getenv('DB_USER', 'joon'),
+        'PASSWORD': os.getenv('DB_PASSWORD', ''),
+        'HOST': os.getenv('DB_HOST', 'localhost'),
+        'PORT': os.getenv('DB_PORT', '5432'),
     }
 }
