+++ manage.py
@@ -7,6 +7,12 @@ import sys
 def main():
     """Run administrative tasks."""
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings.base')
+    
+    # app 폴더를 Python path에 추가
+    app_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'app')
+    if app_path not in sys.path:
+        sys.path.insert(0, app_path)
+    
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
