+++ pyproject.toml
@@ -11,6 +11,9 @@ dependencies = [
     "drf-yasg>=1.21.10",
     "pre-commit>=4.3.0",
     "psycopg2-binary>=2.9.10",
+    "pillow>=11.3.0",
+    "python-dotenv>=1.0.0",
+    "djangorestframework>=3.16.1",
 ]
 
 [dependency-groups]
